Big rename to hf-doc-utils
