+++ README.md
@@ -1,4 +1,4 @@
-# doc-builder
+# hf-doc-utils
 
 This is the package we will use to build the documentation of our Hugging Face repos.
