+++ src/hf_doc_utils.egg-info/PKG-INFO
@@ -1,0 +1,25 @@
+Metadata-Version: 2.1
+Name: hf-doc-utils
+Version: 0.0.1.dev0
+Summary: Doc building utility
+Home-page: https://github.com/huggingface/hf-doc-utils
+Author: Hugging Face, Inc.
+Author-email: sylvain@huggingface.co
+License: UNKNOWN
+Keywords: doc documentation doc-builder huggingface hugging face
+Platform: UNKNOWN
+Description-Content-Type: text/markdown
+Provides-Extra: transformers
+Provides-Extra: testing
+Provides-Extra: quality
+Provides-Extra: all
+Provides-Extra: dev
+License-File: LICENSE
+
+# hf-doc-utils
+
+This is the package we will use to build the documentation of our Hugging Face repos.
+
+This is a work in progress!
+
+
