+++ src/hf_doc_utils.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+hf-doc-utils = hf_doc_utils.commands.hf_doc_utils_cli:main
