+++ src/hf_doc_utils.egg-info/requires.txt
@@ -1,0 +1,36 @@
+tqdm
+pyyaml
+packaging
+nbformat
+
+[all]
+pytest
+pytest-xdist
+torch
+transformers
+black~=22.0
+isort>=5.5.4
+flake8>=3.8.3
+
+[dev]
+pytest
+pytest-xdist
+torch
+transformers
+black~=22.0
+isort>=5.5.4
+flake8>=3.8.3
+
+[quality]
+black~=22.0
+isort>=5.5.4
+flake8>=3.8.3
+
+[testing]
+pytest
+pytest-xdist
+torch
+transformers
+
+[transformers]
+transformers[dev]
